requirements.txt
txt
fastapi
uvicorn
pillow
wordcloud
numpy
rembg
python-multipart